refactor: Extract pre-processor module from message-router
Phase 1 of message-router.ts modularization:
- Create routing/pre-processor.ts (~250 lines)
- Move preProcessIntent() and analyzeAndRouteMessage()
- Move RoutingDecision type and mightNeedTools()
- Contains all deterministic intent detection logic:
  - NBA queries, weather, time, calculator
  - System info, file read, list directory
  - Clipboard, web search patterns
- Re-export from message-router.ts for backward compatibility
- Build verified, app tested and working
